+++ src/main/java/avlyakulov/timur/dto/UserSession.java
@@ -1,0 +1,2 @@
+package avlyakulov.timur.dto;public class UserSession {
+}
