+++ src/main/java/avlyakulov/timur/exception/UserAlreadyExistException.java
@@ -1,0 +1,2 @@
+package avlyakulov.timur.exception;public class UserAlreadyExistException {
+}
