+++ src/main/java/avlyakulov/timur/servlet/UserCreateServlet.java
@@ -1,0 +1,2 @@
+package avlyakulov.timur.servlet;public class UserCreateServlet {
+}
