+++ src/main/java/avlyakulov/timur/servlet/UserEditServlet.java
@@ -1,0 +1,2 @@
+package avlyakulov.timur.servlet;public class UserEditServlet {
+}
